requirements.txt
aiogram==3.21.0
bs4==0.0.2
python-dotenv==1.0.1
langchain==0.3.27
langchain-community==0.3.27
langchain-openai==0.3.30
chromadb==1.0.16
tiktoken==0.6.0
beautifulsoup4==4.12.3
requests==2.31.0

yandex-cloud-ml-sdk==0.14.0
yandexcloud==0.355.0
